requirements.txt
chainlit
python-dotenv
langchain-google-genai
-e .